GUACAMOLE-221: Expose underlying protocol of a GuacamoleSocket via getProtocol().
